LayoutTests/platform/mac-wk2/http/tests/websocket/tests/hybi/client-close-2-expected.txt

WebSocket: Test client-initiated close with 1000.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
Closed
PASS receivedMessage is undefined.
PASS closeEvent.reason is "close_frame[:2]='\\x88\\x82'"
PASS successfullyParsed is true

TEST COMPLETE
